Extend benefits cookie expiry to 30 days
The cookie which tracks when to refresh benefits cookies is reduced to 1
day, like everywhere else.

This is part of a wider change to benefits cookie expiration to fix an
issue where signed-in users with ad free benefits who didn't visit the
site for time period would sometimes see ads on their first returning
pageview. This change makes this less likely to happen by extending the
expiration of user benefits cookies.

This is due to a race condition between the user benefits refresh and
the ads code. However, we don't want to delay ads until after the user
benefits have been refreshed as that would impact performance. So
instead, extend the expiry of the cookie.

See also:
guardian/frontend#28352
guardian/dotcom-rendering#14810
guardian/gateway#3285