0% AB test for using PayPal Complete Payments Zuora payment method for recurring products (#7402)
Add support for recurring products to use the new PayPal Complete Payments
Zuora payment gateway when we create an order in Zuora from support-workers.

An A/B test is used to control the % of payments sent to the new gateway. It's
a single variant test, currently setup to receive 0% of the traffic (so can
only be manually opted-into). As we gain confidence we'll ramp this percentage
up.

This is part of a wider piece of work to upgrade various PayPal elements
(including the REST API endpoints and JS SDK). As a start, Zuora have advised
us that we can migrate to a new payment method type PayPalCP (PayPal Complete
Payments) in isolation. The fields specified in the request to Zuora (BAID and
email) for PayPalCP are identical to those for PayPalNativeEC, which is why
we're able to make this change in isolation.